# backendRL/src/main/java/org/example/backendrl/config/api-docs.yaml
openapi: 3.1.0
info:
  title: RecordingLogs API
  description: "Music rating and recommendation platform: genres, artists, albums,\
    \ songs, ratings, reviews, album lists, favorites and follows."
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/albums/{id}:
    get:
      tags:
      - Albums
      summary: Get an album by id
      operationId: getAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Album found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
    put:
      tags:
      - Albums
      summary: Update an album
      operationId: updateAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
        required: true
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Album updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
    delete:
      tags:
      - Albums
      summary: Delete an album
      operationId: deleteAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Album deleted
...
  /api/lists/{listId}/albums/{albumId}:
    post:
      tags:
      - Lists
      summary: Add an album to a list
      operationId: addAlbum
      parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumListItemRequest"
        required: true
      responses:
        "404":
          description: List or album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: User is not the owner of the list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Album is already in the list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: "Album added, returns the updated list"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumListResponse"
    delete:
      tags:
      - Lists
      summary: Remove an album from a list
      operationId: removeAlbum
      parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: List not found or album not in list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Album removed
...
  /api/albums:
    get:
      tags:
      - Albums
      summary: Get all albums
      operationId: getAllAlbums
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumResponse"
    post:
      tags:
      - Albums
      summary: Create an album
      operationId: createAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Album created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
...
  /api/albums/{albumId}/artists/{artistId}:
    post:
      tags:
      - Albums
      summary: Add an artist to an album
      operationId: addArtistToAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album or artist not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Artist is already on this album
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Artist added to album
    delete:
      tags:
      - Albums
      summary: Remove an artist from an album
      operationId: removeArtistFromAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Album, artist or link not found"
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Artist removed from album
...
  /api/artists/{artistId}/albums:
    get:
      tags:
      - Artists
      summary: Get all albums of an artist
      operationId: getAlbumsForArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Artist not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
...
  /api/albums/{albumId}/artists:
    get:
      tags:
      - Albums
      summary: Get all artists of an album
      operationId: getArtistsForAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of artists
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistSummaryResponse"
components:
  schemas:
...
    AlbumListItemResponse:
      type: object
      properties:
        albumId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
        position:
          type: integer
          format: int32
    AlbumListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        albums:
          type: array
          items:
            $ref: "#/components/schemas/AlbumListItemResponse"
...
    AlbumRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
      required:
      - title
    AlbumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...
    AlbumListItemRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        position:
          type: integer
          format: int32
          minimum: 1
      required:
      - userId
...
    ArtistSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AlbumSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...